requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-appconf==1.0.6
django-autocomplete-light==3.9.7
django-crispy-forms==2.1
django-environ==0.11.2
django-formtools==2.5.1
django-import-export==3.3.5
django-select2==8.1.2
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
python-decouple==3.8
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
